// front/src/app/models/tablas-valores-items.interface.ts
export interface TablasValoresItems {
    id :number|null;
    id_tabla_valor :number|null;
    valor: string|null;
    estado: string|null;
    descripcion: string|null;
    fecha_creacion: Date|null;
    fecha_modificacion: Date|null;
    id_usuario : number|null;
    id_usuario_modificacion :number|null;
}   